requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
atlassian-python-api
beautifulsoup4
python-dotenv
sentence-transformers
faiss-cpu # Use faiss-gpu if you have a compatible GPU